Se cambia Id facebook por long

+++ BackEndRememberMeApp/src/java/Capa_Datos/AgregarPerfil_Datos.java
@@ -18,7 +18,7 @@ public class AgregarPerfil_Datos {
     public AgregarPerfil_Datos() {
     }
 
-    public boolean AgregarPerfil(Perfil perfil, int idFacebook) {
+    public boolean AgregarPerfil(Perfil perfil, long idFacebook) {
         Conexion conexion = new Conexion();
         if (conexion.crearConexion()) {
             PrepararQuerrys preparar = new PrepararQuerrys();
