+++ BackEndRememberMeApp/src/java/Capa_Presentacion/AgregarPerfil_Presentacion.java
@@ -17,7 +17,7 @@ public class AgregarPerfil_Presentacion {
     public AgregarPerfil_Presentacion() {
     }
     
-    public String AgregarPerfil(String jsonPerfil, int idFacebook){
+    public String AgregarPerfil(String jsonPerfil, long idFacebook){
         AgregarPerfil_Logica agregar = new AgregarPerfil_Logica();
         return agregar.AgregarPerfil(jsonPerfil, idFacebook);
     }
